consolidate edit profile and account setup
