+++ frontend/src/assets/sass/_variables.scss
@@ -59,7 +59,7 @@ $primary: $red;
 $secondary: $indigo;
 $success: $green;
 $info: $cyan !default;
-$warning: $yellow;
+$warning: $yellow-300;
 $danger: $red;
 $light: $white;
 $dark: $black;
