+++ frontend/src/assets/sass/lib/_buttons.scss
@@ -102,3 +102,7 @@ button:focus {
     top: 1rem;
   }
 }
+
+.btn-warning.btn-transparent {
+  color: $white;
+}
