+++ frontend/src/components/pages/entertainer/accountSetup/AccountSetup.jsx
@@ -3,13 +3,14 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import TopMessage from 'components/common/layout/TopMessage';
 import BackEndPage from 'components/common/layout/BackEndPage';
+import YoutubeChannelForm from 'components/pages/entertainer/accountSetup/YoutubeChannelForm';
+import IdentificationForm from 'components/pages/entertainer/accountSetup/IdentificationForm';
+import EntertainerDetailsForm from 'components/pages/entertainer/accountSetup/EntertainerDetailsForm';
 import {
-  EntertainerDetailsForm,
-  IdentificationForm,
-  YoutubeChannelForm
-} from 'components/pages/entertainer/EditProfile';
-import { ProfessionalContactForm } from 'components/pages/entertainer/EmergencyContacts';
-import { BankDetailsForm } from 'components/pages/entertainer/BankDetails';
+  ProfessionalContactForm,
+  NextOfKinForm
+} from 'components/pages/entertainer/accountSetup/EmergencyContacts';
+import { BankDetailsForm } from 'components/pages/entertainer/accountSetup/BankDetails';
 import { UserContext } from 'context/UserContext';
 import { ONBOARDING_STEPS } from 'utils/constants';
 
@@ -31,7 +32,10 @@ const AccountSetup = ({ stepFromURL }) => {
     null,
     <EntertainerDetailsForm />,
     <BankDetailsForm />,
-    <ProfessionalContactForm />,
+    <>
+      <ProfessionalContactForm />
+      <NextOfKinForm />
+    </>,
     <YoutubeChannelForm />,
     <IdentificationForm />
   ];
@@ -49,7 +53,7 @@ const AccountSetup = ({ stepFromURL }) => {
   return (
     <BackEndPage title="Account Setup">
       <div className="main-app">
-        <TopMessage message="Account Setup" />
+        <TopMessage message={`Account Setup (Step ${currentStep}/5)`} />
         <section className="app-content">
           <StepperList
             currentStep={currentStep}
@@ -188,3 +192,18 @@ StepperNavigation.propTypes = {
 };
 
 export default AccountSetup;
+
+// TODO:
+// 0. Consolidate Edit profile and account setup, remove unneccessary step (DONE)
+// 1. Add a to complete this section and other forms
+// 2. Update backend to handle updates after approval. [Needs review instead of yes]
+// 3. backend in approval comment, check when everything has been approved and send notification and update approval in entetainers profiile
+// 4. Change sidebar for unapproved users
+// 5. Create videos table
+// 6. post to video endpoint
+// 7. account - create approve all endpoint
+// 8. Recommend a friend, api, facebook, twitter and linkedin
+// 9. Contact us
+// 10. Entertainers full page
+// 11. Switch to user - how do I persist this (use sidebar, check if current route is user route, check to user sidebar)
+// 12. Add a get started button that will create entertainer profile for user
